refactor: rename viz.metrics and model.classification_metrics
