+++ her2bdl/models/__init__.py
@@ -58,7 +58,7 @@ Models and Aggregations
 """
 from .mcdropout import *
 from .uncertainty import *
-from .classification_metrics import *
+from .metrics import *
 
 MODELS = {
     "SimpleClassifierMCDropout" : SimpleClassifierMCDropout,
